4 - Basic String Support (#5)
* Create functions for handling unsigned to signed conversion

* Fixstr, str8, str16, str32

- adds a get_str_type function for grabbing what kind of pack will be used depending on string length
- reworks getsize interface to use int64 instead of default kind
- fixes bugs in NFI packing
- updates a lot of testing

+++ src/messagepack_pack.f90
@@ -17,7 +17,7 @@ module messagepack_pack
             class(mp_value_type) :: mpv
             logical, intent(out) :: error
             byte, allocatable, dimension(:), intent(out) :: buffer
-            integer dblen
+            integer(kind=int64) dblen
 
             call mpv%getsize(dblen) ! get buffer size required
             allocate(buffer(dblen)) ! allocate buffer
